Split up jinja function and add/reconfigure tests
