+++ config/settings/features.py
@@ -1,6 +1,6 @@
 import os
 
-from config.util import strtobool
+from config.utils.env_vars import strtobool
 
 FEATURE_PHASE_BANNER: bool = strtobool(os.getenv("FEATURE_PHASE_BANNER", "True"))
