+++ config/settings/staging.py
@@ -1,6 +1,6 @@
 import os
 
-from config.util import strtobool
+from config.utils.env_vars import strtobool
 
 from .features import *
 from .production import *
