+++ config/utils/number.py
@@ -1,0 +1,6 @@
+def format_number(num):
+    try:
+        number = int(num)
+    except ValueError:
+        return num
+    return format(number, ",")
