+++ test/records/test_archon_xsl_transformations.py
@@ -2,7 +2,7 @@ from django.test import SimpleTestCase
 
 from app.records.constants import TNA_ARCHON_CODE, RecordTypes
 from app.records.models import Record
-from config.jinja import sanitise_record_field
+from config.utils.html import sanitise_record_field
 
 
 class NonTnaArchonRecordTransformationTests(SimpleTestCase):
