+++ test/records/test_related_content.py
@@ -5,8 +5,8 @@ Tests for subjects enrichment functionality
 from unittest.mock import patch
 
 from django.test import TestCase, override_settings
-from django.utils.text import slugify
 from jinja2 import BaseLoader, Environment
+from tna_utilities.string import slugify
 
 from app.lib.exceptions import APIResourceNotFound
 from app.records.api import (  # CHANGED: from views to api
