+++ test/records/test_series_transformations.py
@@ -2,7 +2,7 @@ from django.test import SimpleTestCase
 
 from app.lib.xslt_transformations import SERIES_TRANSFORMATIONS
 from app.records.models import Record
-from config.jinja import sanitise_record_field
+from config.utils.html import sanitise_record_field
 
 
 class SeriesTransformationTests(SimpleTestCase):
